Initial wrapper around AcquireCredentialsHandle
Adds function to call AcquireCredentialsHandleA function with a supplied SSPI package name and return the resultant pointer.

TODO: Figure out why the linker only picks up references to Secur32 when it's placed in the bindings file manually, instead of the references printed in the build script and placed in the .toml file.

TODO: Figure out safer reference to return than mutable pointer.

TODO: Figure out how to convert supplied string to UTF-16 CString so that I can call the W function instead of the A function.